+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - App will now report more strictly errors if you are using a non-compatible source (please report any false positive)
 - It's now possible to always being able to select “None” as an option for air quality, pollen, nowcasting, alerts, normals sources.
 - Fix normals not working in some cases
+- “Location access is off” error introduced in v5.3 will no longer be reported when running in background
 
 **Weather sources**
 - [AccuWeather] When alerts by coordinates is unavailable, it will now fallback to alerts for your city instead of failing completely